tests/requirements_test.txt
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-homeassistant-custom-component>=0.12.0
pytest-timeout>=2.1.0
coverage>=6.0
asynctest>=0.13.0
mock>=4.0.0
numpy>=1.21.0
av>=9.2.0 